requirements.txt
streamlit
openai
python-dotenv
sqlalchemy
pypdf
python-docx
pytesseract
pdf2image
pillow
